+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 
 -   `isOpen` parameter for state `onInit`.
+-   Responsive `data-open` for each size: `data-md-open` ...
+-   `dropy.config.set` object for set dropy config.
 
 ## [v1.2.0](https://github.com/amirHossein5/dropy/compare/v1.1.0...v1.2.0)